# optimize _single_shortest_path function