@dschult 2023-01-09
Does changing the default `join` function to `operator.add` (or even `list.__add__`) speed this up?