@Tortar 2023-01-09
list.__add__ has no effect but with operator.add the function is around 10% faster