@Tortar 2023-05-28
Hi @boothby don't know if you missed this one, but I think it should be also related the optimization made in other PRs you reviewed, what do you think of this one? :-)